// src/app/models/student.model.ts
export class Student {
  _id?: string;
  name?: string;
  email?: string;
  cohort?: string;
  phone_number?: string;

  // _id: string;
  // name: string;
  // email: string;
  // cohort: string;
  // phone_number: string;

  // constructor(
  //   id?: string,
  //   name?: string,
  //   email?: string,
  //   cohort?: string,
  //   phone_number?: string
  // ) {
  //   this._id = id!;
  //   this.name = name!;
  //   this.email = email!;
  //   this.cohort = cohort!;
  //   this.phone_number = phone_number!;
  // }
}